+++ spec/forms/transfer/create_project_form_spec.rb
@@ -299,6 +299,44 @@ RSpec.describe Transfer::CreateProjectForm, type: :model do
         expect(Transfer::TasksData.first.financial_safeguarding_governance_issues).to be true
         expect(Transfer::TasksData.first.outgoing_trust_to_close).to be true
       end
+
+      context "when the project is being assigned to the Regional Caseworker Team" do
+        it "sends a notification to team leaders" do
+          team_leader = create(:user, :team_leader)
+          another_team_leader = create(:user, :team_leader)
+
+          project = build(:create_transfer_project_form, assigned_to_regional_caseworker_team: true).save
+
+          expect(ActionMailer::MailDeliveryJob)
+            .to(have_been_enqueued.on_queue("default")
+                                  .with("TeamLeaderMailer", "new_transfer_project_created", "deliver_now", args: [team_leader, project]))
+          expect(ActionMailer::MailDeliveryJob)
+            .to(have_been_enqueued.on_queue("default")
+                                  .with("TeamLeaderMailer", "new_transfer_project_created", "deliver_now", args: [another_team_leader, project]))
+        end
+
+        context "if a team leader is deactivated" do
+          it "does not send a notification to that team leader" do
+            team_leader = create(:user, :team_leader, deactivated_at: Date.yesterday)
+
+            project = build(:create_transfer_project_form, assigned_to_regional_caseworker_team: true).save
+            expect(ActionMailer::MailDeliveryJob)
+              .to_not(have_been_enqueued.on_queue("default")
+                                        .with("TeamLeaderMailer", "new_transfer_project_created", "deliver_now", args: [team_leader, project]))
+          end
+        end
+      end
+
+      context "when the project is NOT being assigned to the Regional Caseworker Team" do
+        it "does not send a notification to team leaders" do
+          _team_leader = create(:user, :team_leader)
+          _another_team_leader = create(:user, :team_leader)
+
+          _project = build(:create_transfer_project_form, assigned_to_regional_caseworker_team: false).save
+
+          expect(ActionMailer::MailDeliveryJob).to_not(have_been_enqueued)
+        end
+      end
     end
 
     context "when the form is invalid" do